requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
python-dotenv==1.1.1
pydantic==2.12.0
SQLAlchemy==2.0.44
mysql-connector-python==9.5.0
databases==0.9.0
passlib[bcrypt]
bcrypt==4.0.1
python-jose==3.5.0
email-validator==2.2.0
